\name{ComparaNumeros}
\alias{ComparaNumeros}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Funcion que calcula el orden de varios numeros difusos trapezoidales para un metodo.}
\description{
Funcion que calcula el orden de varios numeros difusos trapezoidales para un metodo.
}
\usage{
ComparaNumeros()
}

\details{
Los valores a,b,c,d que se introducen por pantalla tienen que estar ordenados en forma creciente
}
\value{
Devuelve el orden
}
\references{}
\author{msmaldonado}
\note{}
\seealso{}
\examples{
ComparaNumeros()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{RankingNumerosDifusosTrapezoidales}
